A child of a NoopSpan is a NoopSpan. If a NoopSpan or Noop SpanContext is declared as a parent to the span during the SpanBuilder usage, then return the NoopSpanBuilder.
Merge impl and impl-java8 to avoid confusion regarding their roles.
Move Noop classes to their own component "opentracing-noop" where they can't be confused to be part of, or mix-and-matched with, the impl
classes.
Move the Noop classes into interfaces, while creating default implementations of them. This (along with using the create() factory method
name) indicates that this is a marker interface and the instance only should not be compared against.

Opentracing-impl subclasses (extends) the Noop interfaces to provide an empty (or None) instance to the abstract classes provided. This
prevents mix-and-match. Such an approach can be repeated by specific implementations too if they like.

reference: https://github.com/opentracing/opentracing-java/pull/58