Add MDC support for basic ThreadLocalActiveSpan
The documentation heavily mentions MDC in the interfaces but fails to
provide an example in the basic implementations of
ThreadLocalActiveSpan.

This commit adds this functionality via SLF4J which is a very popular
facade to many common logging libraries. As a result, a new dependency
is introduced on slf4j-api however I believe this to be a useful
dependency so that logging can be applied in a runtime agnostic fashion
if additional example implementations would like to be provided.