Add ScopeManager#activate(SpanContext)
Similar to ScopeManager#activate(Span) but used in cases where the
thread in which the activation is performed does not have control over
the life cycle of the span.

This puts the burden of being aware of the lifecycle to the one writing
the instrumentation (for example
io.opentracing.contrib.concurrent.TracedRunnable)
instead of the user of Tracer.activeSpan().

It also simplifies tracer implementations as with this change tracers
don't have to guard against racy calls to Span#finish() and
ScopeManager#activate(Span).