docs(Tracer): remove link to nonexistent method
The previously referred method was removed from the API, hence it no longer exists. Must say too that semantic versioning was used wrong! Removing this method was a breaking change. The same as removing `ScopeManager.activate(Span span, boolean finishSpanOnClose)`. Please be more careful with the versioning of the libs, deprecating a method is better than removing directly :)